+++ Netable/Netable/Netable.swift
@@ -101,7 +101,7 @@ open class Netable {
             }
 
             if T.Parameters.self != Empty.self {
-                try urlRequest.encodeParameters(for: request)
+                try urlRequest.encodeParameters(for: request, defaultEncodingStrategy: config.jsonEncodingStrategy)
             }
         } catch {
             let netableError = (error as? NetableError) ?? NetableError.unknownError(error)
@@ -123,10 +123,16 @@ open class Netable {
         let requestInfo = LogEvent.RequestInfo(
             urlString:  urlRequest.url?.absoluteString ?? "UNDEFINED",
             method: request.method,
-            headers: urlRequest.allHTTPHeaderFields ?? [:],
-            params: try? request.parameters.toParameterDictionary(encodingStrategy: request.jsonKeyEncodingStrategy))
+            headers: urlRequest.allHTTPHeaderFields ?? [:]
+        )
 
         log(.requestStarted(request: requestInfo))
+        if !config.enableLogRedaction, let params = try? request.parameters.toParameterDictionary(encodingStrategy: request.jsonKeyEncodingStrategy ?? config.jsonEncodingStrategy) {
+            log(.requestBody(body: params))
+        } else {
+            let params = request.unredactedParameters(defaultEncodingStrategy: config.jsonEncodingStrategy)
+            log(.requestBody(body: params))
+        }
 
         let retryConfiguration = self.retryConfiguration
 
@@ -144,7 +150,7 @@ open class Netable {
                     throw NetableError.httpError(response.statusCode, data)
                 }
 
-                let decoded = request.decode(data)
+                let decoded = request.decode(data, defaultDecodingStrategy: self.config.jsonDecodingStrategy)
                 switch decoded {
                 case .success(let raw):
                     let finalizedData = request.finalize(raw: raw)