+++ Netable/NetableExample/Requests/LoginRequest.swift
@@ -32,4 +32,8 @@ struct LoginRequest: Request {
    }
 
    public var parameters: LoginParams
+
+    public var unredactedParameterKeys: Set<String> {
+        ["first_name", "username"]
+    }
 }