+++ CHANGELOG.md
@@ -6,8 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [UNRELEASED]
 ### Added
+- [60] Add better handling for printing request bodies to the console, including some default redaction for sensitive parameters.
 - [74] Change Netable constructor to remove the option to directly set a URLSessionConfiguration and instead only expose the `timeout` option.
+- [76] Add an option to set json en/decoding strategies in the Netable config constructor in addition to per-request.
 - [79] Add `requestFailedDelegate` and `requestFailedPublisher` to users to handle errors globally in addition to in `request` completion callbacks. Bumps minimum iOS version to 13.0.
+
 
 ## [0.10.3] - 12-01-21
 ### Changed